Refactor : 파일 저장 관련 로직 변경

+++ src/main/java/teamkiim/koffeechat/domain/file/service/s3/S3PostFileService.java
@@ -1,6 +1,10 @@
 package teamkiim.koffeechat.domain.file.service.s3;
 
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,14 +21,11 @@ import teamkiim.koffeechat.domain.post.common.repository.PostRepository;
 import teamkiim.koffeechat.global.exception.CustomException;
 import teamkiim.koffeechat.global.exception.ErrorCode;
 
-import java.util.List;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 @Profile("prod")
+@Slf4j
 public class S3PostFileService implements PostFileService {
 
     private final FileRepository fileRepository;
@@ -76,18 +77,21 @@ public class S3PostFileService implements PostFileService {
     /**
      * 이미지 파일 다건 삭제 (post에 연관된 File 중 id값이 fileIdList에 없는 File 삭제)
      *
-     * @param fileIdList 삭제하지 않을 이미지 파일 id 리스트
-     * @param post       연관 게시물
-     * @return
+     * @param fileUrlList 삭제하지 않을 이미지 파일 id 리스트
+     * @param post        연관 게시물
      */
     @Transactional
-    public void deleteImageFiles(List<Long> fileIdList, Post post) {
+    public void deleteImageFiles(List<String> fileUrlList, Post post) {
 
         List<PostFile> existFileList = postFileRepository.findAllByPost(post);
 
         List<File> deleteFileList = existFileList.stream()
-                .filter(file -> !fileIdList.contains(file.getId()))
+                .filter(file -> !fileUrlList.contains(file.getUrl()))
                 .collect(Collectors.toList());
+
+        if (!deleteFileList.isEmpty()) {
+            return;
+        }
 
         fileStorageService.deleteFiles(deleteFileList);
